refactor(api): move improve-competence-evaluation and related service
