+++ api/src/evaluation/domain/usecases/improve-competence-evaluation.js
@@ -1,6 +1,6 @@
-import { Assessment } from '../../../src/shared/domain/models/Assessment.js';
-import { MAX_REACHABLE_LEVEL } from '../constants.js';
-import { ImproveCompetenceEvaluationForbiddenError } from '../errors.js';
+import { Assessment } from '../../../shared/domain/models/Assessment.js';
+import { MAX_REACHABLE_LEVEL } from '../../../../lib/domain/constants.js';
+import { ImproveCompetenceEvaluationForbiddenError } from '../../../../lib/domain/errors.js';
 
 const improveCompetenceEvaluation = async function ({
   competenceEvaluationRepository,
