+++ api/src/evaluation/domain/usecases/index.js
@@ -7,6 +7,7 @@ import * as answerRepository from '../../infrastructure/repositories/answer-repo
 import * as assessmentRepository from '../../../shared/infrastructure/repositories/assessment-repository.js';
 import * as competenceRepository from '../../../shared/infrastructure/repositories/competence-repository.js';
 import * as competenceEvaluationRepository from '../../infrastructure/repositories/competence-evaluation-repository.js';
+import { getCompetenceLevel } from '../services/get-competence-level.js';
 
 const path = dirname(fileURLToPath(import.meta.url));
 
@@ -19,6 +20,7 @@ const dependencies = {
   assessmentRepository,
   competenceEvaluationRepository,
   competenceRepository,
+  getCompetenceLevel,
 };
 
 const evaluationUsecases = injectDependencies(usecasesWithoutInjectedDependencies, dependencies);
