+++ api/tests/evaluation/integration/infrastructure/repositories/competence-evaluation-repository_test.js
@@ -1,11 +1,11 @@
 import _ from 'lodash';
 
-import { NotFoundError } from '../../../../../lib/domain/errors.js';
 import { Assessment } from '../../../../../src/shared/domain/models/Assessment.js';
 import { CompetenceEvaluation } from '../../../../../src/evaluation/domain/models/CompetenceEvaluation.js';
 import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
 import * as competenceEvaluationRepository from '../../../../../src/evaluation/infrastructure/repositories/competence-evaluation-repository.js';
 import { catchErr, databaseBuilder, expect, knex } from '../../../../test-helper.js';
+import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
 
 describe('Integration | Repository | Competence Evaluation', function () {
   const STARTED = 'started';
