+++ api/tests/evaluation/unit/domain/services/get-competence-level_test.js
@@ -1,5 +1,5 @@
-import { expect, sinon } from '../../../test-helper.js';
-import { getCompetenceLevel } from '../../../../lib/domain/services/get-competence-level.js';
+import { expect, sinon } from '../../../../test-helper.js';
+import { getCompetenceLevel } from '../../../../../src/evaluation/domain/services/get-competence-level.js';
 
 describe('Unit | Domain | Service | Get Competence Level', function () {
   describe('#getCompetenceLevel', function () {
