+++ api/tests/integration/domain/usecases/improve-competence-evaluation_test.js
@@ -1,9 +1,9 @@
 import { expect, databaseBuilder, knex } from '../../../test-helper.js';
-import { improveCompetenceEvaluation } from '../../../../lib/domain/usecases/improve-competence-evaluation.js';
+import { improveCompetenceEvaluation } from '../../../../src/evaluation/domain/usecases/improve-competence-evaluation.js';
 
 import * as competenceEvaluationRepository from '../../../../src/evaluation/infrastructure/repositories/competence-evaluation-repository.js';
 import * as assessmentRepository from '../../../../src/shared/infrastructure/repositories/assessment-repository.js';
-import { getCompetenceLevel } from '../../../../lib/domain/services/get-competence-level.js';
+import { getCompetenceLevel } from '../../../../src/evaluation/domain/services/get-competence-level.js';
 
 describe('Integration | UseCase | Improve Competence Evaluation', function () {
   const competenceId = 'recCompetenceId';
