+++ api/tests/unit/domain/usecases/improve-competence-evaluation_test.js
@@ -1,5 +1,5 @@
 import { expect, sinon, domainBuilder, catchErr } from '../../../test-helper.js';
-import { improveCompetenceEvaluation } from '../../../../lib/domain/usecases/improve-competence-evaluation.js';
+import { improveCompetenceEvaluation } from '../../../../src/evaluation/domain/usecases/improve-competence-evaluation.js';
 import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
 import { MAX_REACHABLE_LEVEL } from '../../../../lib/domain/constants.js';
 import { ImproveCompetenceEvaluationForbiddenError } from '../../../../lib/domain/errors.js';
